+++ user-ops-indexer/user-ops-indexer-logic/src/types/bundle.rs
@@ -6,7 +6,7 @@ use ethers_core::{abi::AbiEncode, utils::to_checksum};
 #[derive(Clone, Debug, PartialEq)]
 pub struct Bundle {
     pub transaction_hash: H256,
-    pub bundle_index: u64,
+    pub bundle_index: u32,
     pub block_number: u64,
     pub bundler: Address,
     pub timestamp: String,
@@ -17,7 +17,7 @@ impl From<BundleDB> for Bundle {
     fn from(v: BundleDB) -> Self {
         Self {
             transaction_hash: H256::from_slice(&v.transaction_hash),
-            bundle_index: v.bundle_index as u64,
+            bundle_index: v.bundle_index as u32,
             block_number: v.block_number as u64,
             bundler: Address::from_slice(&v.bundler),
             total_ops: v.total_ops as u32,
